Make PackagesInDir ignore errors on excluded paths
Fixes issue where "PackagesInDir" would fail if the executing
user could not read a file or directory in the provided rootDir,
even if the file or directory was excluded by the exclude matcher.

Cannot add test because there is no way to reliably simulate an
unreadable file/directory in a test.